+++ ostap/fitting/tests/test_fitting_roostats2.py
@@ -53,42 +53,38 @@ model.S = 55
 model.B = 1000
 model.S.setMax(200)
 
-data    = model.generate ( 55 + 1000 )
-data1   = data.clone ()
-data2   = data.clone ()
-data3   = data.clone ()
-data4   = data.clone ()
-data5   = data.clone ()
-data6   = data.clone ()
+data = model.generate ( 55 + 1000 )
+
+summary = [ ('method' , '90%CL' , 'time [s]') ]
+plots   = []
 
 # ============================================================================-
 ## Get the upper limit limit for small signal at fixed mass
 #  - resolution is fixed
 #  - Asymptotic Calculator is used 
-def test_point_limit_ac () :
+def test_point_limit_ac1() :
     """Get the upper limit at given point for small signal at fixed mass
     - resoltuion is fixed 
     - Asymptotic Calculator is used 
     """
 
-    logger = getLogger("test_point_limit_ac")
+    logger = getLogger("test_point_limit_ac1")
 
-    logger.info ( "Test Point limits with RooStats (Asymptotic Calcultor)" )
+    logger.info ( "Test Point limits with RooStats (Asymptotic Calculator)" )
 
     from   ostap.fitting.roostats   import ( ModelConfig           ,
                                              AsymptoticCalculator  ,
                                              HypoTestInverter      )
 
     the_model = model.clone ( name = 'M1' )
-
-    with use_canvas ( 'test_point_limit_ac' ) : 
-        logger.info ( 'Dataset is\n%s' % data1.table ( prefix = '# ' ) ) 
-        rr , frame = the_model.fitTo ( data1 , draw = True , nbins = 50 )
+    
+    with use_canvas ( 'test_point_limit_ac1' ) : 
+        rr , frame = the_model.fitTo ( data , draw = True , nbins = 50 )
 
     ## create ModelConfig  for 'S+B' model
     model_sb = ModelConfig ( pdf       = the_model   ,
                              poi       = the_model.S , ## parameter of interest 
-                             dataset   = data1       ,
+                             dataset   = data        ,
                              name      = 'S+B'       )
     
     model_sb.snapshot = the_model.S ## ATTENTION! 
@@ -96,7 +92,7 @@ def test_point_limit_ac () :
     ## create ModelConfig  for 'B-only' model
     model_b  = ModelConfig ( pdf       = the_model          ,
                              poi       = the_model.S        , ## parameter of interest 
-                             dataset   = data1              ,
+                             dataset   = data               ,
                              workspace = model_sb.workspace , 
                              name      = 'B-only'           )
     
@@ -106,11 +102,11 @@ def test_point_limit_ac () :
     logger.info ( 'Model config %s\n%s'  % ( model_sb.name , model_sb.table ( prefix = '# ' ) ) ) 
     logger.info ( 'Model config %s\n%s'  % ( model_b.name  , model_b .table ( prefix = '# ' ) ) )
     
-    with timing ( "Using Asymptotic Calculator" , logger = logger ) :
+    with timing ( "Using Asymptotic Calculator" , logger = logger ) as timer :
         ## create the calculator 
         ac  = AsymptoticCalculator ( model_b           ,
                                      model_sb          ,
-                                     dataset   = data1 ,
+                                     dataset   = data  ,
                                      asimov    = False )
         ac.calculator.SetOneSided ( True ) 
         
@@ -118,15 +114,97 @@ def test_point_limit_ac () :
         hti = HypoTestInverter ( ac ,  0.90 , use_CLs = True , verbose = False )
         
         ## make a scan 
-        hti .scan ( vrange ( 0 , 150 , 50 )  ) ## scan it!
+        hti .scan_with_progress ( vrange ( 0 , 150 , 150 )  ) ## scan it!
         
     ## visualize the scan results 
-    with use_canvas ( 'test_pointLimit: HypoTestInverter plot (asymptotic)' , wait = 2 ) :
+    with use_canvas ( 'test_point_limit_ac1: HypoTestInverter plot (asymptotic)' , wait = 2 ) :
         plot = hti.plot
         plot .draw('LCb 2CL')                    
         logger.info ( '90%%CL upper limit (asymptotic)  = %.1f' % hti.upper_limit )
 
-    logger.info ( 'Dataset is\n%s' % data1.table ( prefix = '# ' ) ) 
+        row = 'Asymptotic' , '%.1f' % hti.upper_limit , '%.1f' % timer.delta 
+        summary.append ( row  )
+        plots  .append ( plot )
+        
+    ## check the dataset
+    stat = data.statVar('mass')
+    if stat.rms() <= 0 :
+        logger.error   ( 'Calculator destroyed input dataset!') 
+        logger.error   ( 'Dataset is\n%s'          % data.table ( prefix = '# ' ) ) 
+
+# ============================================================================-
+## Get the upper limit limit for small signal at fixed mass
+#  - resolution is fixed
+#  - Asymptotic Calculator is used 
+def test_point_limit_ac2() :
+    """Get the upper limit at given point for small signal at fixed mass
+    - resoltuion is fixed 
+    - Asymptotic Calculator is used 
+    """
+
+    logger = getLogger("test_point_limit_ac2")
+
+    logger.info ( "Test Point limits with RooStats (Asymptotic Calculator with Asimov dataset)" )
+
+    from   ostap.fitting.roostats   import ( ModelConfig           ,
+                                             AsymptoticCalculator  ,
+                                             HypoTestInverter      )
+
+    the_model = model.clone ( name = 'M2' )
+    
+    with use_canvas ( 'test_point_limit_ac2' ) : 
+        rr , frame = the_model.fitTo ( data , draw = True , nbins = 50 )
+
+    ## create ModelConfig  for 'S+B' model
+    model_sb = ModelConfig ( pdf       = the_model   ,
+                             poi       = the_model.S , ## parameter of interest 
+                             dataset   = data        ,
+                             name      = 'S+B'       )
+    
+    model_sb.snapshot = the_model.S ## ATTENTION! 
+    
+    ## create ModelConfig  for 'B-only' model
+    model_b  = ModelConfig ( pdf       = the_model          ,
+                             poi       = the_model.S        , ## parameter of interest 
+                             dataset   = data               ,
+                             workspace = model_sb.workspace , 
+                             name      = 'B-only'           )
+    
+    the_model.S = 0 
+    model_b.snapshot = the_model.S  ## ATTENTION! 
+    
+    logger.info ( 'Model config %s\n%s'  % ( model_sb.name , model_sb.table ( prefix = '# ' ) ) ) 
+    logger.info ( 'Model config %s\n%s'  % ( model_b.name  , model_b .table ( prefix = '# ' ) ) )
+    
+    with timing ( "Using Asymptotic/Asimov Calculator" , logger = logger ) as timer :
+        ## create the calculator 
+        ac  = AsymptoticCalculator ( model_b           ,
+                                     model_sb          ,
+                                     dataset   = data  ,
+                                     asimov    = True  )
+        ac.calculator.SetOneSided ( True ) 
+        
+        ## create Hypo Test inverter 
+        hti = HypoTestInverter ( ac ,  0.90 , use_CLs = True , verbose = False )
+        
+        ## make a scan 
+        hti .scan_with_progress ( vrange ( 0 , 150 , 150 )  ) ## scan it!
+        
+    ## visualize the scan results 
+    with use_canvas ( 'test_point_limit_ac2: HypoTestInverter plot (asymptotic/asimov)' , wait = 2 ) :
+        plot = hti.plot
+        plot .draw('LCb 2CL')                    
+        logger.info ( '90%%CL upper limit (asymptotic/asimov)  = %.1f' % hti.upper_limit )
+
+        row = 'Asymptotic/Asimov' , '%.1f' % hti.upper_limit, '%.1f' % timer.delta 
+        summary.append ( row  )
+        plots  .append ( plot )
+        
+    ## check the dataset
+    stat = data.statVar('mass')
+    if stat.rms() <= 0 :
+        logger.error   ( 'Calculator destroyed input dataset!') 
+        logger.error   ( 'Dataset is\n%s'          % data.table ( prefix = '# ' ) ) 
 
 # ============================================================================-
 ## Get the upper limit limit for small signal at fixed mass
@@ -150,16 +228,15 @@ def test_point_limit_fc  () :
                                              FrequentistCalculator ,
                                              HypoTestInverter      )
 
-    the_model = model.clone ( name = 'M2' )
-
+    the_model = model.clone ( name = 'M3' )
+    
     with use_canvas ( 'test_point_limit_fc' ) : 
-        logger.info ( 'Dataset is\n%s' % data2.table ( prefix = '# ' ) ) 
-        rr , frame = the_model.fitTo ( data2 , draw = True , nbins = 50 )
+        rr , frame = the_model.fitTo ( data , draw = True , nbins = 50 )
 
     ## create ModelConfig  for 'S+B' model
     model_sb = ModelConfig ( pdf       = the_model   ,
                              poi       = the_model.S , ## parameter of interest 
-                             dataset   = data2       ,
+                             dataset   = data        ,
                              name      = 'S+B'       )
     
     model_sb.snapshot = the_model.S ## ATTENTION! 
@@ -167,7 +244,7 @@ def test_point_limit_fc  () :
     ## create ModelConfig  for 'B-only' model
     model_b  = ModelConfig ( pdf       = the_model          ,
                              poi       = the_model.S        , ## parameter of interest 
-                             dataset   = data2              ,
+                             dataset   = data               ,
                              workspace = model_sb.workspace , 
                              name      = 'B-only'           )
     
@@ -178,31 +255,37 @@ def test_point_limit_fc  () :
     logger.info ( 'Model config %s\n%s'  % ( model_b.name  , model_b .table ( prefix = '# ' ) ) )
     
     ## with Frequentist calculator
-    with timing ( "Using Frequentist Calculator" , logger = logger ) :
+    with timing ( "Using Frequentist Calculator" , logger = logger ) as timer :
         
         ## create the calculator 
         fc  = FrequentistCalculator ( model_b            ,
                                       model_sb           ,
-                                      dataset    = data2 ,
-                                      ntoys_null = 50    ,
-                                      ntoys_alt  = 50    ,
+                                      dataset    = data  ,
+                                      ntoys_null = 100   ,
+                                      ntoys_alt  = 100   ,
                                       ) 
-        
         
         ## create Hypo Test inverter 
         hti = HypoTestInverter ( fc ,  0.90 , use_CLs = True , verbose = False )
         
         ## make a scan 
-        hti .scan_with_progress  ( vrange ( 0 , 100 , 20 )  ) ## scan it!
+        hti .scan_with_progress  ( vrange ( 0.1 , 100 , 10 )  ) ## scan it!
  
     ## visualize the scan results 
-    with use_canvas ( 'test_pointLimit: HypoTestInverter plot (frequentist)' , wait = 2 ) :
+    with use_canvas ( 'test_point_limits_fc: HypoTestInverter plot (frequentist)' , wait = 2 ) :
         plot = hti .plot
         plot .draw('LCb 2CL')    
         logger.info ( '90%%CL upper limit (frequentist) = %.1f' % hti.upper_limit )
 
-    logger.warning ( 'Frequenstist cclaultor destgroys input dataset') 
-    logger.info    ( 'Dataset is\n%s' % data2.table ( prefix = '# ' ) ) 
+        row = 'Frequentist' , '%.1f' % hti.upper_limit, '%.1f' % timer.delta
+        summary.append ( row  )
+        plots  .append ( plot )
+
+    ## check the dataset
+    stat = data.statVar('mass')
+    if stat.rms() <= 0 :
+        logger.error   ( 'Calculator destroyed input dataset!') 
+        logger.error   ( 'Dataset is\n%s'          % data.table ( prefix = '# ' ) ) 
 
 # ============================================================================-
 ## Get the upper limit limit for small signal at fixed mass
@@ -218,24 +301,19 @@ def test_point_limit_hc  () :
 
     logger.info ( "Test Point limits with RooStats using Hybrid Calculator" )
 
-    ## if root_info < (6,24) :
-    ##    logger.info ( 'Test is disabled for ROOT version %s' % str ( root_info ) )
-    ##    return 
-
     from   ostap.fitting.roostats   import ( ModelConfig           ,
                                              HybridCalculator      ,
                                              HypoTestInverter      )
 
-    the_model = model.clone ( name = 'M3' )
-
+    the_model = model.clone ( name = 'M4' )
+    
     with use_canvas ( 'test_point_limit_hc' ) : 
-        logger.info ( 'Dataset is\n%s' % data3.table ( prefix = '# ' ) ) 
-        rr , frame = the_model.fitTo ( data3 , draw = True , nbins = 50 )
+        rr , frame = the_model.fitTo ( data , draw = True , nbins = 50 )
 
     ## create ModelConfig  for 'S+B' model
     model_sb = ModelConfig ( pdf       = the_model   ,
                              poi       = the_model.S , ## parameter of interest 
-                             dataset   = data3       ,
+                             dataset   = data        ,
                              name      = 'S+B'       )
     
     model_sb.snapshot = the_model.S ## ATTENTION! 
@@ -243,7 +321,7 @@ def test_point_limit_hc  () :
     ## create ModelConfig  for 'B-only' model
     model_b  = ModelConfig ( pdf       = the_model          ,
                              poi       = the_model.S        , ## parameter of interest 
-                             dataset   = data3              ,
+                             dataset   = data               ,
                              workspace = model_sb.workspace , 
                              name      = 'B-only'           )
     
@@ -254,14 +332,14 @@ def test_point_limit_hc  () :
     logger.info ( 'Model config %s\n%s'  % ( model_b.name  , model_b .table ( prefix = '# ' ) ) )
     
     ## with Hybrid calculator
-    with timing ( "Using Hybrid Calculator" , logger = logger ) :
+    with timing ( "Using Hybrid Calculator" , logger = logger ) as timer :
         
         ## create the calculator 
         hc  = HybridCalculator ( model_b            ,
                                  model_sb           ,
-                                 dataset    = data3 ,
-                                 ntoys_null = 50    ,
-                                 ntoys_alt  = 50    ,
+                                 dataset    = data  ,
+                                 ntoys_null = 100   ,
+                                 ntoys_alt  = 100   ,
                                  ) 
         
         
@@ -269,16 +347,23 @@ def test_point_limit_hc  () :
         hti = HypoTestInverter ( hc ,  0.90 , use_CLs = True , verbose = False )
         
         ## make a scan 
-        hti .scan_with_progress  ( vrange ( 0 , 100 , 20 )  ) ## scan it!
+        hti .scan_with_progress  ( vrange ( 0.1 , 100 , 10 )  ) ## scan it!
  
     ## visualize the scan results 
-    with use_canvas ( 'test_pointLimit: HypoTestInverter plot (hybrid)' , wait = 2 ) :
+    with use_canvas ( 'test_point_limit_hc: HypoTestInverter plot (hybrid)' , wait = 2 ) :
         plot = hti .plot
         plot .draw('LCb 2CL')    
         logger.info ( '90%%CL upper limit (hybrid) = %.1f' % hti.upper_limit )
 
-    logger.info ( 'Dataset is\n%s' % data3.table ( prefix = '# ' ) ) 
+        row = 'Hybrid' , '%.1f' % hti.upper_limit, '%.1f' % timer.delta
+        summary.append ( row  )
+        plots  .append ( plot )
 
+    ## check the dataset
+    stat = data.statVar('mass')
+    if stat.rms() <= 0 :
+        logger.error   ( 'Calculator destroyed input dataset!') 
+        logger.error   ( 'Dataset is\n%s'          % data.table ( prefix = '# ' ) ) 
 
 # ============================================================================-
 ## Get the upper limit limit for small signal at fixed mass
@@ -298,16 +383,15 @@ def test_point_limit_pl () :
                                              ProfileLikelihoodCalculator ,
                                              HypoTestInverter            )
 
-    the_model = model.clone ( name = 'M1' )
-
+    the_model = model.clone ( name = 'M5' )
+    
     with use_canvas ( 'test_point_limit_ac' ) : 
-        logger.info ( 'Dataset is\n%s' % data4.table ( prefix = '# ' ) ) 
-        rr , frame = the_model.fitTo ( data4 , draw = True , nbins = 50 )
+        rr , frame = the_model.fitTo ( data , draw = True , nbins = 50 )
 
     ## create ModelConfig  for 'S+B' model
     model_sb = ModelConfig ( pdf       = the_model   ,
                              poi       = the_model.S , ## parameter of interest 
-                             dataset   = data4       ,
+                             dataset   = data        ,
                              name      = 'S+B'       )
     
     model_sb.snapshot = the_model.S ## ATTENTION! 
@@ -315,7 +399,7 @@ def test_point_limit_pl () :
     ## create ModelConfig  for 'B-only' model
     model_b  = ModelConfig ( pdf       = the_model          ,
                              poi       = the_model.S        , ## parameter of interest 
-                             dataset   = data4              ,
+                             dataset   = data               ,
                              workspace = model_sb.workspace , 
                              name      = 'B-only'           )
     
@@ -368,22 +452,21 @@ def test_point_limit2 () :
     sigma = ROOT.RooRealVar( 'sigma_Gau1', 'sigma of Gaussian' , 0.3 , 0.1 , 2 )
 
     the_signal = signal.clone ( sigma = sigma , name = 'S1' )
-
+    
     ## create "soft" constraint for sigma 
     sigma_constraint = the_signal.soft_constraint ( sigma , VE ( 0.3 , 0.01**2 ) ) 
     
-    the_model = model.clone ( name = 'M5' , signal = the_signal , signals = () )
+    the_model = model.clone ( name = 'M6' , signal = the_signal , signals = () )
 
     ## all constraints 
     constraints = sigma_constraint, 
 
     with use_canvas ( 'test_point_limit2' ) : 
-        logger.info ( 'Dataset is\n%s' % data5.table ( prefix = '# ' ) ) 
-        rr , frame = the_model.fitTo ( data5 , draw = True , nbins = 50 , constraints = constraints )
+        rr , frame = the_model.fitTo ( data , draw = True , nbins = 50 , constraints = constraints )
         
     model_sb = ModelConfig ( pdf         = the_model        ,
                              poi         = the_model.S      , ## parameter of interest 
-                             dataset     = data5            ,
+                             dataset     = data             ,
                              constraints = constraints      ,   
                              name        = 'S+B'            )
     
@@ -392,7 +475,7 @@ def test_point_limit2 () :
     
     model_b  = ModelConfig ( pdf         = the_model          ,
                              poi         = the_model.S        , ## parameter of interest 
-                             dataset     = data5              ,
+                             dataset     = data               ,
                              workspace   = model_sb.workspace , 
                              constraints = constraints        ,   
                              name        = 'B-only'           )
@@ -403,22 +486,31 @@ def test_point_limit2 () :
     logger.info ( 'Model config %s\n%s'  % ( model_sb.name , model_sb.table ( prefix = '# ' ) ) ) 
     logger.info ( 'Model config %s\n%s'  % ( model_b.name  , model_b .table ( prefix = '# ' ) ) )
 
-    ac  = AsymptoticCalculator ( model_b           ,
-                                 model_sb          ,
-                                 dataset   = data5 )
-    ac.calculator.SetOneSided ( True ) 
+    with timing ( "Using Asymptotic Calculator with 1 constraint" , logger = logger ) as timer :
+        
+        ac  = AsymptoticCalculator ( model_b           ,
+                                     model_sb          ,
+                                     dataset   = data  )
+        ac.calculator.SetOneSided ( True ) 
 
-    hti = HypoTestInverter ( ac ,  0.90 , use_CLs = True , verbose = False )
+        hti = HypoTestInverter ( ac ,  0.90 , use_CLs = True , verbose = False )
     
-    hti.scan ( vrange ( 0 , 150 , 50 )  ) ## scan it!
+        hti.scan_with_progress ( vrange ( 0 , 150 , 150 )  ) ## scan it!
 
     with use_canvas ( 'test_point_limit2: HypoTestInverter plot' , wait = 2 ) :
         plot = hti.plot
         plot.draw('LCb 2CL')    
         logger.info ( '90%%CL upper limit = %.1f' % hti.upper_limit )
+        plots  .append ( plot )
 
-    logger.info ( 'Dataset is\n%s' % data5.table ( prefix = '# ' ) ) 
+        row = 'Asymptotic (1 constraint)' , '%.1f' % hti.upper_limit , '%.1f' % timer.delta 
+        summary.append ( row  )
 
+    ## check the dataset
+    stat = data.statVar('mass')
+    if stat.rms() <= 0 :
+        logger.error   ( 'Calculator destroyed input dataset!') 
+        logger.error   ( 'Dataset is\n%s'          % data.table ( prefix = '# ' ) ) 
 
 # ============================================================================-
 ## Get the upper limit limit for small signal at fixed mass 
@@ -440,7 +532,7 @@ def test_point_limit3 () :
     sigma = ROOT.RooRealVar( 'sigma_Gau2', 'sigma of Gaussian' , 0.3 , 0.1 , 2 )
 
     the_signal = signal.clone ( sigma = sigma , name = 'S2' )
-
+    
     ## create "soft" constraint for sigma: (0.30+/-0.01)  
     sigma_constraint = the_signal.soft_constraint ( sigma , VE ( 0.3 , 0.01**2 ) ) 
 
@@ -456,17 +548,16 @@ def test_point_limit3 () :
     ## raw/visible signal yield 
     raw_S = the_signal.vars_multiply ( NS , eff , 'raw_S' , 'raw/observed signal yeild' )
 
-    the_model = model.clone ( name = 'M6' , signal = the_signal , signals = () , S = raw_S )
+    the_model = model.clone ( name = 'M7' , signal = the_signal , signals = () , S = raw_S )
 
     constraints = sigma_constraint, eff_constraint 
 
     with use_canvas ( 'test_point_limit3' ) : 
-        logger.info ( 'Dataset is\n%s' % data6.table ( prefix = '# ' ) ) 
-        rr , frame = the_model.fitTo ( data6 , draw = True , nbins = 50 , constraints = constraints )
+        rr , frame = the_model.fitTo ( data , draw = True , nbins = 50 , constraints = constraints )
         
     model_sb = ModelConfig ( pdf         = the_model        ,
                              poi         = NS               , ## parameter of interest 
-                             dataset     = data6            ,
+                             dataset     = data             ,
                              constraints = constraints      ,   
                              name        = 'S+B'            )
     
@@ -475,7 +566,7 @@ def test_point_limit3 () :
     
     model_b  = ModelConfig ( pdf         = the_model          ,
                              poi         = NS                 , ## parameter of interest 
-                             dataset     = data6              ,
+                             dataset     = data               ,
                              workspace   = model_sb.workspace , 
                              constraints = constraints        ,   
                              name        = 'B-only'           )
@@ -486,40 +577,53 @@ def test_point_limit3 () :
     logger.info ( 'Model config %s\n%s'  % ( model_sb.name , model_sb.table ( prefix = '# ' ) ) ) 
     logger.info ( 'Model config %s\n%s'  % ( model_b.name  , model_b .table ( prefix = '# ' ) ) )
 
-    ac  = AsymptoticCalculator ( model_b           ,
-                                 model_sb          ,
-                                 dataset   = data6 )
-    ac.calculator.SetOneSided ( True )
-    
+    with timing ( "Using Asymptotic Calculator with 2 constraints" , logger = logger ) as timer :
 
-    hti = HypoTestInverter ( ac ,  0.90 , use_CLs = True , verbose = False )
-    
-    hti.scan ( vrange ( 0 , 150 , 50 )  ) ## scan it!
+        ac  = AsymptoticCalculator ( model_b           ,
+                                     model_sb          ,
+                                     dataset   = data  )
+        ac.calculator.SetOneSided ( True )
+                
+        hti = HypoTestInverter ( ac ,  0.90 , use_CLs = True , verbose = False )
+        
+        hti.scan_with_progress ( vrange ( 0 , 150 , 150 )  ) ## scan it!
 
     with use_canvas ( 'test_point_limit3: HypoTestInverter plot' , wait = 2 ) :
         plot = hti.plot
         plot.draw('LCb 2CL')    
         logger.info ( '90%%CL upper limit = %.1f' % hti.upper_limit )
+        plots  .append ( plot )
+        row = 'Asymptotic (2 constraints' , '%.1f' % hti.upper_limit , '%.1f' % timer.delta 
+        summary.append ( row  )
 
-    logger.info ( 'Dataset is\n%s' % data6.table ( prefix = '# ' ) ) 
+    ## check the dataset
+    stat = data.statVar('mass')
+    if stat.rms() <= 0 :
+        logger.error   ( 'Calculator destroyed input dataset!') 
+        logger.error   ( 'Dataset is\n%s'          % data.table ( prefix = '# ' ) ) 
 
-    
 # =============================================================================
 if '__main__' == __name__ :
 
     from ostap.core.core    import rooSilent
     
     with rooSilent ( ) : 
-    
-        test_point_limit_ac ()        
-        test_point_limit_fc ()
-        test_point_limit_hc ()
         
-        ## test_point_limit_pl ()
+        test_point_limit_ac1 ()
+        test_point_limit_ac2 ()
+        test_point_limit_fc  ()
+        test_point_limit_hc  ()
         
-        test_point_limit2   ()
-        test_point_limit3   ()
+        ## ## test_point_limit_pl ()
+        
+        test_point_limit2    ()
+        test_point_limit3    ()
 
+    import ostap.logger.table as T
+    title = '90%CL upper limits'
+    table = T.table ( summary , title = title , prefix = '# ' , alignment = 'lcr' )
+    logger.info ( '%s:\n%s' % ( title , table ) )
+    
 # =============================================================================
 ##                                                                      The END 
 # =============================================================================