+++ ReleaseNotes/release_notes.md
@@ -1,6 +1,8 @@
 ## New features:
 
   1. add new example on making p-value scan (thanks to Dima Golubkov) 
+  1. Fix `FrequentistsCalcualtor` and `HybridCalculator` to use cloned datasets. (They can destriy/corrupt input dataset). Clone ddatatsets ar deleted after usage 
+
 
 ## Backward incompatible: